# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TurbineOps REST API
  description: Turbine inspection and maintenance management API
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:4000/api
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Bearer token obtained from /auth/login endpoint'
  schemas:
...
    Inspection:
      type: object
      properties:
        id:
          type: string
        turbineId:
          type: string
        date:
          type: string
          format: date
        dataSource:
          type: string
          enum: [DRONE, MANUAL]
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /inspections:
    get:
      summary: List inspections (optionally filter by turbine)
      tags: [Inspections]
      parameters:
        - name: turbineId
          in: query
          schema:
            type: string
          description: Filter inspections by turbine ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of results to skip
      responses:
        '200':
          description: Inspections list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inspection'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized
    post:
      summary: Create inspection
      tags: [Inspections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - turbineId
                - date
                - dataSource
              properties:
                turbineId:
                  type: string
                date:
                  type: string
                  format: date
                dataSource:
                  type: string
                  enum: [DRONE, MANUAL]
                inspectorName:
                  type: string
                rawPackageUrl:
                  type: string
      responses:
        '201':
          description: Inspection created
        '400':
          description: Invalid request or overlap detected
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role
  /inspections/{id}:
    get:
      summary: Get inspection details
      tags: [Inspections]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inspection details
        '401':
          description: Unauthorized
        '404':
          description: Inspection not found

    put:
      summary: Update inspection
      tags: [Inspections]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                dataSource:
                  type: string
                  enum: [DRONE, MANUAL]
      responses:
        '200':
          description: Inspection updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role

    delete:
      summary: Delete inspection
      tags: [Inspections]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Inspection deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN role
  /inspections/{inspectionId}/findings:
    post:
      summary: Create finding
      tags: [Findings]
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category
                - severity
              properties:
                category:
                  type: string
                  enum: [BLADE_DAMAGE, LIGHTNING, EROSION, UNKNOWN]
                severity:
                  type: integer
                  minimum: 1
                  maximum: 5
                notes:
                  type: string
      responses:
        '201':
          description: Finding created
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role
  /findings/{id}:
    get:
      summary: Get finding details
      tags: [Findings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Finding details
        '401':
          description: Unauthorized

    put:
      summary: Update finding
      tags: [Findings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                  enum: [BLADE_DAMAGE, LIGHTNING, EROSION, UNKNOWN]
                severity:
                  type: integer
                  minimum: 1
                  maximum: 5
                notes:
                  type: string
      responses:
        '200':
          description: Finding updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role

    delete:
      summary: Delete finding
      tags: [Findings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Finding deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN role
...